backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
bcrypt
email-validator
python-multipart
